requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
email-validator
passlib
python-jose[cryptography]
dotenv
logging
bandit